+++ lib/modules/state/state_service.dart
@@ -1,9 +1,9 @@
 ﻿// Module: state
-// Updated at: 2026-07-03 13:13:40
-// Commit: test(state): implement widget integration tests for state
+// Updated at: 2026-07-03 14:26:12
+// Commit: feat(state): add bi-directional event bus handler for state
 
 class stateComponentState {
-  final String commitRef = '2524';
-  final String lastMessage = 'test(state): implement widget integration tests for state';
-  final DateTime timestamp = DateTime.parse('2026-07-03T13:13:40+05:00');
+  final String commitRef = '2532';
+  final String lastMessage = 'feat(state): add bi-directional event bus handler for state';
+  final DateTime timestamp = DateTime.parse('2026-07-03T14:26:12+05:00');
 }